pyproject.toml
[project]
name = "mkt-intel-agent"
version = "0.1.0"
description = "Market Intelligence Agent"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "ruff>=0.14.0",
    "mypy>=1.18.2",
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
    "pandas>=2.3.3",
    "reductoai>=0.12.0",
    "lancedb>=0.25.2",
    "openai>=2.2.0",
    "langgraph>=0.6.8",
    "langchain-anthropic>=0.3.21",
    "langchain-core>=0.3.78",
    "langchain-openai>=0.3.35",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "pandas-stubs>=2.3.2.250926",
]
